Se modificó la clase lectura, se eliminaron algunos metodos y se crearon otros tantos, con el fin de mejorar la organizacion de esta clase.
Se cambió el formato del archivo de entrada.

Se esta modificando la clase heristica, aun no se termina, tiene algunos errores.

+++ Tesis v3.0/src/especialidad.cpp
@@ -21,7 +21,7 @@ Especialidad::~Especialidad() {}
 
 int Especialidad::id(void) const
 {
-	return static_cast<int>(_id);
+	return _id;
 }
 
 string Especialidad::nombre(void) const
